Tests/LibWeb/Text/expected/grid/percentage-grid-item-minimum-contribution.txt

document scroll width: 1280
section grid-template-columns: 1280px
container width: 1100
cards width: 1100
cards scroll width exceeds client width: true
